smiga wszystko raczek zostalo szyfrowanie hasla i cos z ta waga podzialac w month summary

+++ Dietaverse/Dietaverse/Database/db_model.edmx.sql
@@ -2,7 +2,7 @@
 -- --------------------------------------------------
 -- Entity Designer DDL Script for SQL Server 2005, 2008, 2012 and Azure
 -- --------------------------------------------------
--- Date Created: 10/15/2022 15:23:32
+-- Date Created: 10/15/2022 15:45:18
 -- Generated from EDMX file: C:\Users\zuzia\source\repos\Dietaverse\Dietaverse\Dietaverse\Database\db_model.edmx
 -- --------------------------------------------------
 
@@ -77,7 +77,7 @@ CREATE TABLE [dbo].[dishes_gallerySet] (
     [Id] int IDENTITY(1,1) NOT NULL,
     [name] nvarchar(max)  NULL,
     [recipe] nvarchar(max)  NULL,
-    [kcal] float  NULL,
+    [kcal] int  NULL,
     [photo] nvarchar(max)  NOT NULL,
     [users_Id] int  NOT NULL
 );
